fix: Copilotレビュー対応 — isPayloadTooLargeError共通化・警告改善・入力順序修正
- isPayloadTooLargeErrorをfigma-client.tsに共通ヘルパーとして抽出し、
  REST adapterとdesign-checkの両方で再利用
- --input-dirのマーカーをCLI引数順に即時pushするよう修正
- チャンキング警告を多ページ/大ページで条件分岐
